implemented screen navigation with react-navigation, added .eslintrc.json so that we can all code a little cleaner
