+++ app/.eslintrc.json
@@ -1,0 +1,52 @@
+{
+	"plugins": [
+		"react",
+		"react-native"
+	],
+  "env": {
+      "browser": true,
+      "es6": true,
+      "node": true,
+			"react-native/react-native": true
+  },
+  "extends": ["eslint:recommended", "plugin:react-native/all"],
+  "parserOptions": {
+    "ecmaFeatures": {
+          "experimentalObjectRestSpread": true,
+					"jsx": true
+    },
+		"sourceType": "module"
+  },
+  "rules": {
+    "semi": [
+        "error",
+        "always"
+    ],
+    "no-var": [
+        "error"
+    ],
+    "prefer-const": ["error", {
+        "destructuring": "any",
+        "ignoreReadBeforeAssign": false
+    }],
+    "curly": ["error"],
+    "eqeqeq": ["error"],
+    "no-multi-spaces": ["error"],
+    "no-lone-blocks": ["error"],
+    "no-self-compare": ["error"],
+    "no-unused-expressions": ["error"],
+    "no-useless-call": ["error"],
+    "no-use-before-define": ["error"],
+
+    "camelcase": ["error", {"properties": "never"}],
+    "func-call-spacing": ["error"],
+    "no-lonely-if": ["error"],
+    "array-bracket-spacing": ["error"],
+
+    "no-console": ["off"],
+
+    "no-unused-vars": ["error", { "args": "none" }],
+
+		"react/jsx-uses-vars": ["error"]
+  }
+}
