+++ app/Signup.js
@@ -1,4 +1,3 @@
-Signup.js
 
 import React from 'react';
 import { StyleSheet, Text, View, TextInput, Image, TouchableOpacity } from 'react-native';
