+++ app/package.json
@@ -3,8 +3,11 @@
   "version": "0.1.0",
   "private": true,
   "devDependencies": {
-    "react-native-scripts": "1.11.1",
+    "eslint": "^4.18.1",
+    "eslint-plugin-react": "^7.7.0",
+    "eslint-plugin-react-native": "^3.2.1",
     "jest-expo": "25.0.0",
+    "react-native-scripts": "1.11.1",
     "react-test-renderer": "16.2.0"
   },
   "main": "./node_modules/react-native-scripts/build/bin/crna-entry.js",
@@ -27,6 +30,6 @@
     "react-native": "0.52.0",
     "react-native-elements": "^0.19.0",
     "react-native-material-design": "^0.3.7",
-    "react-router-native": "^4.2.0"
+    "react-navigation": "^1.3.0"
   }
 }
